Update Kotlin and IntelliJ dependency
This one took a lot of effort. Turns out that for some reason on Windows the tests thought they were running in a headless environment, so I had to manually set headless to false in build.gradle.

I had to rewrite DataInsertActionTest completely because apparently the actions can no longer be mocked.

Additionally, I fixed a few things for #209.