Remove unnecessary information from preview errors
The additional information was introduced by accident in #376 when implementing #373 to add a timeout for generating values. Error messages in the preview (and other places) would contain the full name of the exception that caused the error, which is completely unnecessary.

This fix was originally implemented in the `uds` branch, but I decided to add a "backports" branch for fixes I might want to release in a v2.x.x release before releasing v3.0.0.