Remove scheme panel and isolate primitive schemes
Decimals, integers, etc. can no longer be edited directly from the main menu, and their code has had all the boilerplate removed so that they are now "isolated" units that can be embedded into the UDS dialog when the context requires it. I'm not quite sure what to do with arrays, so for now they'll just behave like the UDS dialog.