Overhaul UDS structure to use list of primitives
The goal is now to use JSON to serialise the primitives safely back and forth between its representation of an array of schemes which are individually configured. The interface will consist first of a list of available templates (currently not implemented), and then the list of schemes the template consists of in the next list. Finally, a panel from which the currently-selected scheme can be edited. The second-level list can be used by users to add new schemes, including another template because another template is also a scheme.

A future goal is also to allow users to specify how arrays and repetitions and so on should work together when inserted.